fix markdown route
